refactor(rust/sedona-functions): move ST_Transform and sd_order out of sedona-proj
Relocate the ST_Transform and sd_order (OrderLngLat) CRS functions from
sedona-proj to sedona-functions, and move the SedonaOptions -> runtime CRS
engine resolution (with_crs_engine) into sedona-common, where its config-less
fallback is the erroring DefaultCrsEngine rather than the global PROJ engine.
The session registers LazyProjEngine as the default, so real invocations are
unaffected; reprojecting tests and the benchmark inject LazyProjEngine.

The concrete PROJ engine (ProjCrsEngine/LazyProjEngine/with_global_proj_engine)
stays in sedona-proj, which no longer hosts DataFusion UDFs and drops its
datafusion-expr, sedona-functions, sedona-expr, and arrow dependencies.
