+++ c/sedona-proj/src/lib.rs
@@ -19,6 +19,4 @@ pub mod error;
 mod proj;
 mod proj_dyn_bindgen;
 pub mod register;
-pub mod sd_order_lnglat;
-mod st_transform;
 pub mod transform;
