+++ rust/sedona-functions/src/lib.rs
@@ -19,6 +19,7 @@ pub mod executor;
 pub mod register;
 mod sd_format;
 pub mod sd_order;
+pub mod sd_order_lnglat;
 mod sd_simplifystorage;
 mod st_affine;
 mod st_affine_helpers;
@@ -61,6 +62,7 @@ pub mod st_setsrid;
 mod st_srid;
 mod st_start_point;
 mod st_togeomgeog;
+mod st_transform;
 mod st_translate;
 mod st_xyzm;
 mod st_xyzm_minmax;
